requirements.txt
automium==0.2.6
automium-web==0.1.1
click==6.7
codegen==1.0
Django==2.0.6
djangorestframework==3.8.2
Flask==1.0.2
itsdangerous==0.24
Jinja2==2.10
legacy==0.1.6
MarkupSafe==1.0
mysqlclient==1.3.12
patterns==0.3
pytz==2018.4
views==0.3
Werkzeug==0.14.1
